Update README.md
AtliQ Hardware is growing rapidly in the recent years, and they have decided to implement the data analytics using PowerBI in their company for the first time to make data driven decisions & surpass their competitors. This project is to give answers to the questions of stakeholder in terms all the aspects like Finance, Sales, Marketing and Supply Chain.
By using these insights, decisions can be taken based on the data. It will further help in answering numerous questions based on the situation.